Extracted error InternalError to a separate file.
