+++ packages/ckeditor5-dev-changelog/src/errors/InternalError.ts
