+++ packages/ckeditor5-dev-changelog/src/utils/getsectionstodisplay.ts
@@ -4,7 +4,8 @@
  */
 
 import type { Section, SectionsWithEntries } from '../types.js';
-import { InternalError } from '../errors.js';
+
+import { InternalError } from '../errors/InternalError.js';
 
 /**
  * This function determines which sections contain valid entries to be shown.
